refactor: move MessagesList and BadWordsMessage to seperate components
